[Feat] timeout용 API 응답을 RecordingResultsResponse로 수정 - 다음 루트 질문을 띄워줄 수 있도록 함
