+++ src/main/java/com/dgu/review/domain/interview/service/RecordingQueryService.java
@@ -1,7 +1,5 @@
 package com.dgu.review.domain.interview.service;
 
-import com.dgu.review.domain.interview.dto.response.NextPayload;
-import com.dgu.review.domain.interview.dto.response.NextQuestionType;
 import com.dgu.review.domain.interview.dto.response.ProgressStatus;
 import com.dgu.review.domain.interview.dto.response.RecordingResultsResponse;
 import com.dgu.review.domain.interview.entity.InterviewQuestion;
@@ -22,13 +20,13 @@ public class RecordingQueryService {
 
     private final RecordingRepository recordingRepo;
     private final RecordingStatusService statusService;
+    private final NextQuestionPlanner nextQuestionPlanner;
 
     public RecordingResultsResponse getRecordingResults(Long recordingId) {
         var recording = recordingRepo.findById(recordingId)
                 .orElseThrow(() -> new ApiException(ErrorCode.RECORDING_NOT_FOUND));
 
         InterviewQuestion current = recording.getInterviewQuestion();
-        InterviewQuestion root = getRoot(current);
         var status = statusService.getStatus(recordingId);
 
         if (status == RecordingStatus.FAILED || recording.getFailedAt() != null) {
@@ -43,7 +41,7 @@ public class RecordingQueryService {
                 (status == null && recording.getFailedAt() == null);
 
         if (isSuccess) {
-            var next = decideNextPayload(current, root);
+            var next = nextQuestionPlanner.decideNextPayload(current);
             return RecordingResultsResponse.builder()
                     .sessionId(current.getInterviewSession().getId())
                     .status(ProgressStatus.READY)
@@ -57,66 +55,6 @@ public class RecordingQueryService {
                 .next(null)
                 .build();
 
-    }
-
-    private NextPayload decideNextPayload(InterviewQuestion current, InterviewQuestion currentRoot) {
-        var savedFollowUp = current.getFollowUpQuestion();
-
-        // 꼬리질문 존재
-        if (savedFollowUp != null) {
-            return NextPayload.builder()
-                    .type(NextQuestionType.FOLLOW_UP)
-                    .nextQuestionId(savedFollowUp.getId())
-                    .nextQuestionText(savedFollowUp.getQuestion())
-                    .rootId(currentRoot.getId())
-                    .rootText(currentRoot.getQuestion())
-                    .rootIndex(currentRoot.getQuestionNumber())
-                    .build();
-        }
-
-        // 꼬리질문 x -> 다음 루트 질문
-        var nextRoot = findNextRoot(currentRoot);
-        if (nextRoot != null) {
-            return NextPayload.builder()
-                    .type(NextQuestionType.ROOT)
-                    .nextQuestionId(nextRoot.getId())
-                    .nextQuestionText(nextRoot.getQuestion())
-                    .rootId(nextRoot.getId())
-                    .rootText(nextRoot.getQuestion())
-                    .rootIndex(nextRoot.getQuestionNumber())
-                    .build();
-        }
-
-        // 더 이상 낼 질문이 없으면 NONE - 세션 종료
-        return NextPayload.builder()
-                .type(NextQuestionType.NONE)
-                .rootId(currentRoot.getId())
-                .rootText(currentRoot.getQuestion())
-                .rootIndex(currentRoot.getQuestionNumber())
-                .build();
-    }
-
-    private InterviewQuestion findNextRoot(InterviewQuestion root) {
-        return root.getInterviewSession().getQuestions().stream()
-                .filter(q -> q.getParentQuestion() == null)
-                .filter(q -> q.getQuestionNumber() == root.getQuestionNumber() + 1)
-                .findFirst()
-                .orElse(null);
-    }
-
-    private InterviewQuestion getRoot(InterviewQuestion q) {
-        var cur = q;
-        int depth = 0;
-
-        while (cur.getParentQuestion() != null) {
-            cur = cur.getParentQuestion();
-
-            if (++depth > 10) {
-                log.error("질문 데이터 순환 참조 의심. recordingId: {}", q.getRecording().getId());
-                throw new ApiException(ErrorCode.DATA_INTEGRITY_VIOLATED);
-            }
-        }
-        return cur;
     }
 
 }
